backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2026.1.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.20.0",
]
